Move copying logic to separate class
