+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/KlumInstanceProxy.java
@@ -38,7 +38,6 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 import static com.blackbuild.klum.ast.util.DslHelper.*;
-import static groovyjarjarasm.asm.Opcodes.*;
 import static java.lang.String.format;
 
 /**
@@ -157,7 +156,7 @@ public class KlumInstanceProxy {
      * @param template The template to apply
      */
     public void copyFrom(Object template) {
-        DslHelper.getDslHierarchyOf(instance.getClass()).forEach(it -> copyFromLayer(it, template));
+        new CopyHandler(instance).copyFrom(template);
     }
 
     public Object cloneInstance() {
@@ -165,41 +164,6 @@ public class KlumInstanceProxy {
         Object result = FactoryHelper.createInstance(instance.getClass(), (String) key);
         getProxyFor(result).copyFrom(instance);
         return result;
-    }
-
-    private void copyFromLayer(Class<?> layer, Object template) {
-        if (layer.isInstance(template))
-            Arrays.stream(layer.getDeclaredFields())
-                    .filter(this::isNotIgnored)
-                    .forEach(field -> copyFromField(field, template));
-    }
-
-    private boolean isIgnored(Field field) {
-        if ((field.getModifiers() & (ACC_SYNTHETIC | ACC_FINAL | ACC_TRANSIENT)) != 0) return true;
-        if (field.isAnnotationPresent(Key.class)) return true;
-        if (field.isAnnotationPresent(Owner.class)) return true;
-        if (field.isAnnotationPresent(Role.class)) return true;
-        if (field.getName().startsWith("$")) return true;
-        if (DslHelper.getKlumFieldType(field) == FieldType.TRANSIENT) return true;
-        return false;
-    }
-
-    private boolean isNotIgnored(Field field) {
-        return !isIgnored(field);
-    }
-
-    private void copyFromField(Field field, Object template) {
-        String fieldName = field.getName();
-        Object templateValue = getProxyFor(template).getInstanceAttribute(fieldName);
-
-        if (templateValue == null) return;
-
-        if (templateValue instanceof Collection)
-            copyFromCollectionField((Collection<?>) templateValue, fieldName);
-        else if (templateValue instanceof Map)
-            copyFromMapField((Map<?,?>) templateValue, fieldName);
-        else
-            setInstanceAttribute(fieldName, getCopiedValue(templateValue));
     }
 
     @SuppressWarnings("unchecked")
