+++ klum-ast-runtime/src/test/groovy/com/blackbuild/klum/ast/util/KlumInstanceProxyTest.groovy
@@ -24,7 +24,7 @@
 //file:noinspection GrPackage
 package com.blackbuild.klum.ast.util
 
-import spock.lang.Issue
+
 import spock.lang.Subject
 
 class KlumInstanceProxyTest extends AbstractRuntimeTest {
@@ -146,158 +146,6 @@ class KlumInstanceProxyTest extends AbstractRuntimeTest {
         !proxy.resolveKeyForFieldFromAnnotation("noFieldAnnotation", proxy.getField("noFieldAnnotation")).isPresent()
     }
 
-    @Issue("36")
-    def "copy from creates copies of nested DSL objects"() {
-        given:
-        createClass('''
-            package pk
-
-import com.blackbuild.groovy.configdsl.transform.DSL
-
-            @SuppressWarnings('UnnecessaryQualifiedReference')
-            @DSL
-            class Outer {
-                KlumInstanceProxy $proxy = new KlumInstanceProxy(this)
-                String name
-                
-                Inner inner
-            }
-            
-            @DSL
-            class Inner {
-                KlumInstanceProxy $proxy = new KlumInstanceProxy(this)
-                String value
-            } 
-         ''')
-
-        def inner = newInstanceOf("pk.Inner")
-        def outer = newInstanceOf("pk.Outer")
-        inner.value = "bla"
-        outer.inner = inner
-        outer.name = "bli"
-
-        when:
-        def copy = newInstanceOf("pk.Outer")
-        proxy = KlumInstanceProxy.getProxyFor(copy)
-        proxy.copyFrom(outer)
-
-        then:
-        copy.name == "bli"
-        copy.inner.value == "bla"
-        !copy.inner.is(inner)
-    }
-
-    @SuppressWarnings('GroovyAssignabilityCheck')
-    @Issue("36")
-    def "copy from creates copies of nested DSL object collections and maps"() {
-        given:
-        createClass('''
-            package pk
-
-import com.blackbuild.groovy.configdsl.transform.DSL
-
-            @SuppressWarnings('UnnecessaryQualifiedReference')
-            @DSL
-            class Outer {
-                KlumInstanceProxy $proxy = new KlumInstanceProxy(this)
-                String name
-                
-                List<Inner> inners = []
-                Map<String, Inner> mappedInners = [:]
-            }
-            
-            @DSL
-            class Inner {
-                KlumInstanceProxy $proxy = new KlumInstanceProxy(this)
-                String value
-            } 
-         ''')
-
-        def inner = newInstanceOf("pk.Inner")
-        def inner2 = newInstanceOf("pk.Inner")
-        def minner = newInstanceOf("pk.Inner")
-        def minner2 = newInstanceOf("pk.Inner")
-        def outer = newInstanceOf("pk.Outer")
-        inner.value = "bla"
-        inner2.value = "blu"
-        minner.value = "mbla"
-        minner2.value = "mblu"
-        outer.inners.add inner
-        outer.inners.add inner2
-
-        outer.mappedInners.putAll(one: minner, two: minner2)
-        outer.name = "bli"
-
-        when:
-        def copy = newInstanceOf("pk.Outer")
-        proxy = KlumInstanceProxy.getProxyFor(copy)
-        proxy.copyFrom(outer)
-
-        then:
-        copy.name == "bli"
-        !copy.inners.is(outer.inners)
-        copy.inners.size() == 2
-
-        and:
-        copy.inners[0].value == "bla"
-        copy.inners[1].value == "blu"
-        !copy.inners[0].is(inner)
-        !copy.inners[1].is(inner2)
-
-        and:
-        copy.mappedInners.one.value == "mbla"
-        copy.mappedInners.two.value == "mblu"
-        !copy.mappedInners.one.is(minner)
-        !copy.mappedInners.two.is(minner2)
-    }
-
-    @Issue("36")
-    def "copy from creates copies of Maps of Lists"() {
-        given:
-        createClass('''
-            package pk
-
-            import com.blackbuild.groovy.configdsl.transform.DSL
-
-            @SuppressWarnings('UnnecessaryQualifiedReference')
-            @DSL
-            class Outer {
-                KlumInstanceProxy $proxy = new KlumInstanceProxy(this)
-                String name
-                
-                Map<String, List<String>> inners = [:]
-                List<List<String>> innerLists = []
-            }
-         ''')
-
-        def outer = newInstanceOf("pk.Outer")
-        outer.name = "bli"
-        outer.inners.put "a", ["a1", "a2"]
-        outer.inners.put "b", ["b1", "b2"]
-        outer.innerLists.add(["a1", "a2"])
-        outer.innerLists.add(["b1", "b2"])
-
-        when:
-        def copy = newInstanceOf("pk.Outer")
-        proxy = KlumInstanceProxy.getProxyFor(copy)
-        proxy.copyFrom(outer)
-
-        then:
-        copy.name == "bli"
-        copy.inners == outer.inners
-        !copy.inners.is(outer.inners)
-        copy.innerLists == outer.innerLists
-        !copy.innerLists.is(outer.innerLists)
-
-        and:
-        copy.inners.a == outer.inners.a
-        !copy.inners.a.is(outer.inners.a)
-
-        and:
-        copy.innerLists[0] == outer.innerLists[0]
-        !copy.innerLists[0].is(outer.innerLists[0])
-
-    }
 
     // TODO: List of Lists, mixed dsl / non dsl elements
 }
